split out OMQueue component into separate files
